Add deprecation notice to README
Hey there,

I tried first to reach out through an issue or an email, but it looks like the issue tracker was closed, and there was no email on your profile, so I thought I'd go with a PR.

I recently joined the [grunt-broccoli](https://github.com/EmberSherpa/grunt-broccoli) project, and just now pushed a version 1.0.0 of that module to npm. It depends on the latest version of Broccoli, cleans up the API a little and added a new feature. To me (who's been using Broccoli outside of ember-cli for several years), there seems to be some confusion around which grunt-broccoli module was the right one to use. There are multiple forks on npm, since grunt-broccoli itself hadn't been updated for some time.

Anyway, I think it would be great to consolidate our efforts in [EmberSherpa/grunt-broccoli](https://github.com/EmberSherpa/grunt-broccoli), and work together to make that module the best it can be!

Best regards,
Fredrik